move code shared between functions and runtime
