+++ functions-v1/tsconfig.json
@@ -14,6 +14,6 @@
   "compileOnSave": true,
   "include": [
     "src",
-    "test"
-  ]
+    "test",
+    "../shared"  ]
 }
