+++ shared/shared-utils.ts
