+++ src/components/document/doc-list-debug.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { IDocumentMetadata } from "../../../functions/src/shared";
+import { IDocumentMetadata } from "../../../shared/shared";
 
 import "./doc-list-debug.scss";
